+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Add file reload retries to (hopefully) fix occassional skipping of game events
 - `ChromaController` will now create a `NativeApi` instance only if no `IChromaApi` instance has been provided
 
+### Added
+
+- Add Steam Libraries to game install folder detection routine
+
 ## [1.1.1](https://github.com/poveden/EliteChroma/compare/v1.1.0...v1.1.1) — 2020-01-30
 
 ### Fixed